MiniV5/requirements_mcbv7.txt
faiss-cpu
numpy
pypdf
sentence-transformers
groq
